Yggio/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "yggio-client"
version = "0.1.0"
description = "Yggio client for communicating with Yggio API and retrieving data"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Forum Virium Helsinki"},
]
keywords = ["yggio", "api", "sensors", "iot", "data-processing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "httpx",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
# Main API clients
yggio-client = "yggio_client:main"
yggio-tool = "yggio_tool:main"

[tool.setuptools]
py-modules = ["yggio_client", "yggio_tool"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["yggio_client"]
